requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-pymongo==2.3.0
pymongo==4.8.0
python-dotenv==1.0.1

ultralytics==8.3.0
opencv-python==4.10.0.84
deepface==0.0.93
numpy==1.26.4
scikit-learn==1.5.1

qrcode[pil]==7.4.2
reportlab==4.2.2

PyJWT==2.9.0
Werkzeug==3.0.3
